requirements.txt
fastapi
uvicorn[standard]
pydantic
google-auth
google-api-python-client
python-dotenv
